requirements.txt
quart==0.17.0
minio==7.1.6
python-dotenv==0.20.0
orjson==3.6.7
uvicorn[standard]==0.17.6
uvloop==0.16.0; platform_system!="Windows" # https://github.com/MagicStack/uvloop/issues/14